pyproject.toml
[project]
name = "chronoreason-kdsh-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.0",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "pathway>=0.27.1",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.10",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
]
